+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Added: The map tracker now supports random elevators, translator gates and starting location.
 
+-   Changed: The pickup spoiler in game details is now sorted.
+
 -   Fixed: Multiworld sessions should no longer occasionally duplicate messages.
 
 ## [2.0.0] - 2020-11-15