+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Fixed: Multiworld sessions should no longer occasionally duplicate messages.
 
+-   Fixed: Custom safe zone healing should now work in multiworld sessions.
+
 ## [2.0.0] - 2020-11-15
 
 This version is dedicated to SpaghettiToastBook, a great member of our community who sadly lost her life this year.