+++ CHANGELOG.md
@@ -14,7 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Fixed: Multiworld sessions should no longer occasionally duplicate messages.
 
+-   Added: Starting locations tab has checkboxes to easily select all locations in an area
+    
 -   Fixed: Custom safe zone healing should now work in multiworld sessions.
+
 
 ## [2.0.0] - 2020-11-15