+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.1] - 2020-11-??
+
 -   Added: The map tracker now supports random elevators, translator gates and starting location.
 
 -   Changed: The pickup spoiler in game details is now sorted.
@@ -17,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added: Starting locations tab has checkboxes to easily select all locations in an area
     
 -   Fixed: Custom safe zone healing should now work in multiworld sessions.
+
+-   Fixed: Occasional error with switching an observer into a player.
 
 
 ## [2.0.0] - 2020-11-15